+++ resources/views/la/accounts/show.blade.php
@@ -8,6 +8,9 @@
 @section('main-content')
 <div id="page-content" class="profile2">
     <div class="bg-primary clearfix">
+        <div class="col-md-12">
+            <h4 class="name">{{  ($account->title_id > 0)  ? $account->title->description : ''  }}  {{ $account->name1 }} {{ $account->name2 }}</h4>
+        </div>
         <div class="col-md-4">
             <div class="row">
                 <div class="col-md-3">
@@ -15,8 +18,8 @@
                     <div class="profile-icon text-primary"><i class="fa {{ $module->fa_icon }}"></i></div>
                 </div>
                 <div class="col-md-9">
-                    <h4 class="name">{{ $account->$view_col }}</h4>
                     <div class="row stats">
+                        <div class="col-md-12" style="background:white">@la_display($module, 'account_account_type','',0)</div>
                         <div class="col-md-4"><i class="fa fa-facebook"></i> 234</div>
                         <div class="col-md-4"><i class="fa fa-twitter"></i> 12</div>
                         <div class="col-md-4"><i class="fa fa-instagram"></i> 89</div>
@@ -26,7 +29,8 @@
             </div>
         </div>
         <div class="col-md-3">
-            <div class="dats1"><div class="label2">Admin</div></div>
+            <div class="dats1"><div class="label2">{{ ($account->is_person == 0) ? 'Società' : 'Persona' }}</div></div>
+
             <div class="dats1"><i class="fa fa-envelope-o"></i> superadmin@gmail.com</div>
             <div class="dats1"><i class="fa fa-map-marker"></i> Pune, India</div>
         </div>
@@ -82,13 +86,18 @@
         </div>
         <div class="col-md-1 actions">
             @la_access("Accounts", "edit")
-                <a href="{{ url(config('laraadmin.adminRoute') . '/accounts/'.$account->id.'/edit') }}" class="btn btn-xs btn-edit btn-default"><i class="fa fa-pencil"></i></a><br>
+                @if ( $account->hasAccess($account->id,'edit', Auth::user()->id) == 1)
+                    <a href="{{ url(config('laraadmin.adminRoute') . '/accounts/'.$account->id.'/edit') }}" class="btn btn-xs btn-edit btn-default"><i class="fa fa-pencil"></i></a><br>
+                @endif
             @endla_access
 
             @la_access("Accounts", "delete")
-                {{ Form::open(['route' => [config('laraadmin.adminRoute') . '.accounts.destroy', $account->id], 'method' => 'delete', 'style'=>'display:inline']) }}
-                    <button class="btn btn-default btn-delete btn-xs" type="submit"><i class="fa fa-times"></i></button>
-                {{ Form::close() }}
+                @if ( $account->hasAccess($account->id,'delete', Auth::user()->id) == 1)
+
+                    {{ Form::open(['route' => [config('laraadmin.adminRoute') . '.accounts.destroy', $account->id], 'method' => 'delete', 'style'=>'display:inline']) }}
+                        <button class="btn btn-default btn-delete btn-xs" type="submit"><i class="fa fa-times"></i></button>
+                    {{ Form::close() }}
+                @endif
             @endla_access
         </div>
     </div>
@@ -96,8 +105,11 @@
     <ul data-toggle="ajax-tab" class="nav nav-tabs profile" role="tablist">
         <li class=""><a href="{{ url(config('laraadmin.adminRoute') . '/accounts') }}" data-toggle="tooltip" data-placement="right" title="Back to Accounts"><i class="fa fa-chevron-left"></i></a></li>
         <li class="active"><a role="tab" data-toggle="tab" class="active" href="#tab-general-info" data-target="#tab-info"><i class="fa fa-bars"></i> General Info</a></li>
+        <li class=""><a role="tab" data-toggle="tab" href="#tab-contacts" data-target="#tab-contacts"><i class="fa fa-user"></i> contacts</a></li>
         <li class=""><a role="tab" data-toggle="tab" href="#tab-timeline" data-target="#tab-timeline"><i class="fa fa-clock-o"></i> Timeline</a></li>
-        <li class=""><a role="tab" data-toggle="tab" href="#tab-access_rights" data-target="#tab-access_rights"><i class="fa fa-key"></i> access_rights</a></li>
+        @role(['SUPER_ADMIN','NORMAL_ADMIN'])
+            <li class=""><a role="tab" data-toggle="tab" href="#tab-access_rights" data-target="#tab-access_rights"><i class="fa fa-key"></i> access_rights</a></li>
+        @endrole
     </ul>
 
     <div class="tab-content">
@@ -108,17 +120,22 @@
                         <h4>General Info</h4>
                     </div>
                     <div class="panel-body">
-                        @la_display($module, 'title_id')
-						@la_display($module, 'is_person')
-						@la_display($module, 'name1')
-						@la_display($module, 'name2')
+                        {{-- @la_display($module, 'title_id') --}}
+						{{-- @la_display($module, 'is_person') --}}
+						{{-- @la_display($module, 'name1') --}}
+						{{-- @la_display($module, 'name2') --}}
 						@la_display($module, 'notes')
-						@la_display($module, 'account_account_type')
-						@la_display($module, 'account_user')
+						{{-- @la_display($module, 'account_account_type') --}}
+                        @la_display($module, 'vat_number')
+                        @la_display($module, 'fiscal_code')
+                        @la_display($module, 'is_blocked')
+                        @la_display($module, 'block_alert_message')
+                    
                     </div>
                 </div>
             </div>
         </div>
+        @include('la.accounts.tabs.contacts')
         @include('la.accounts.tabs.timeline')
         @include('la.accounts.tabs.access_rights')
     </div>
